Convinced the code that actually an underflow is not possible

+++ lib/Basics/StringHeap.cpp
@@ -53,7 +53,8 @@ arangodb::velocypack::StringRef StringHeap::registerString<arangodb::velocypack:
 template <>
 arangodb::velocypack::HashedStringRef StringHeap::registerString<arangodb::velocypack::HashedStringRef>(arangodb::velocypack::HashedStringRef const& value) {
   char const* p = registerString(value.data(), value.size());
-  return arangodb::velocypack::HashedStringRef(p, value.size());
+  // We got a uint32_t size string in, we do not modify it, so static cast here is save.
+  return arangodb::velocypack::HashedStringRef(p, static_cast<uint32_t>(value.size()));
 }
 
 void StringHeap::clear() {
